Task-140426/backend/pyproject.toml
[project]
name = "legal-qa-api"
version = "1.0.0"
description = "Legal Document Q&A API — Traditional RAG + Agentic RAG"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "pymupdf4llm>=0.0.17",
    "pymupdf>=1.24.11",
    "langchain>=0.3.15",
    "langchain-openai>=0.2.8",
    "langchain-qdrant>=0.2.0",
    "langchain-community>=0.3.15",
    "langgraph>=0.2.0",
    "qdrant-client>=1.11.3",
    "pydantic>=2.0.0",
]
